+++ src/mysql-printer.js
@@ -3,7 +3,9 @@ const mysql = require('mysql')
 
 class MySQLPrinter extends SQLPrinter {
   async print (job) {
-    this.tableName = this._name(job.name)
+    // Allow for the table name to be set by environment variable
+    // If not present, uses the job name
+    this.tableName = process.env.MYSQL_TABLE ? process.env.MYSQL_TABLE : this._name(job.name)
     try {
       return await super.print(job, true)
     } catch (e) {
@@ -55,6 +57,9 @@ class MySQLPrinter extends SQLPrinter {
 
   _query (sql, params) {
     return new Promise((resolve, reject) => {
+      if (params) {
+        params = params.map(param => this._value(param))
+      }
       // console.info('SQL: ' + sql)
       const options = {
         sql: sql,
@@ -68,6 +73,12 @@ class MySQLPrinter extends SQLPrinter {
         // results will contain the results of the query
         // fields will contain information about the returned results fields (if any)
         if (error) {
+          if (!sql.includes('CREATE TABLE')) {
+            console.error('MYSQL QUERY error on sql: ' + sql)
+            if (params) {
+              console.error('MYSQL QUERY error on params:' + params.join(',\n'))
+            }
+          }
           reject(error)
         } else {
           resolve(results)